question-answering-app/question-answering-backend/requirements.txt
Flask==3.0.0
openai==0.28.1
pinecone-client==2.2.4
tiktoken==0.5.1
